[UPDATE BACKEND] Eliminar atributo newPath de EditProductParamsDto y corregir método editProduct de ProductManagementService
	- Eliminado warnings de AddProductParamsDto y ProductManagementServiceTest

+++ backend/src/main/java/es/udc/tfgproject/backend/model/services/ProductManagementServiceImpl.java
@@ -71,7 +71,7 @@ public class ProductManagementServiceImpl implements ProductManagementService {
 
 		/* Antes de añadir la nueva foto, hay que eliminar del sistema la anterior */
 
-		if (product.getImage() != null) {
+		if (product.getImage() != null && newPath != null) {
 			Image imageToRemove = imageDao.findByPath(product.getImage().getPath()).get();
 			// deleteFile(imageToRemove.getPath());
 			imageDao.delete(imageToRemove);
